modules/locations-consumer/requirements.txt
GeoAlchemy2==0.9.4
greenlet==1.1.2
kafka-python==2.0.2
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
packaging==21.0
psycopg2-binary==2.9.1
pyparsing==3.0.1
Shapely==1.8.0
six==1.16.0
SQLAlchemy==1.4.26
wheel==0.37.0